crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
supabase==2.0.5
python-dotenv==1.0.0
lxml==4.9.3
pillow==10.1.0
aiohttp==3.9.1
asyncio
fake-useragent==1.4.0